Finally Recovered the entire code after accidentally triggering pull request

+++ src/app/recipe.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Recipe } from './recipe.interface';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
 @Injectable({
@@ -9,20 +9,20 @@ import { Observable } from 'rxjs';
 export class RecipeService {
 
 
-  private baseUrl = 'http://localhost:3000/Recipe';
+  private baseUrl = 'http://localhost:3000/recipes';
 
   constructor(private httpClient: HttpClient) {}
 
   getAllRecipe(): Observable<Recipe[]> {
-    return this.httpClient.get<Recipe[]>(this.baseUrl);
+    return this.httpClient.get<Recipe[]>(this.baseUrl,{ headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   addRecipe(Recipe: Recipe) {
-    return this.httpClient.post(this.baseUrl,Recipe);
+    return this.httpClient.post(this.baseUrl,Recipe,{ headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   deleteRecipe(id: string) {
-    return this.httpClient.delete(this.baseUrl + '/' + id);
+    return this.httpClient.delete(this.baseUrl + '/' + id,{ headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
